Reinstate setting of ssh->session_started.
When PuTTY is configured to display stderr diagnostics from the proxy
command but only until the main session starts, this flag is how the
SSH backend indicates the point at which the session starts. It was
previously set during version-string parsing, and I forgot to find a
new home for it when I moved the version string parsing out into the
new verstring BPP module in commit af8e526a7.

Now reinstated, at the point where that BPP gets back to us and tells
us what protocol version it's chosen.